+++ src/transformers/models/blip/image_processing_pil_blip.py
@@ -13,13 +13,13 @@
 # limitations under the License.
 """Image processor class for BLIP."""
 
-from ...image_processing_backends import TorchvisionBackend
+from ...image_processing_backends import PilBackend
 from ...image_utils import OPENAI_CLIP_MEAN, OPENAI_CLIP_STD, PILImageResampling
 from ...utils import auto_docstring
 
 
 @auto_docstring
-class BlipImageProcessor(TorchvisionBackend):
+class BlipImageProcessorPil(PilBackend):
     resample = PILImageResampling.BICUBIC
     image_mean = OPENAI_CLIP_MEAN
     image_std = OPENAI_CLIP_STD
@@ -31,4 +31,4 @@ class BlipImageProcessor(TorchvisionBackend):
     do_convert_rgb = True
 
 
-__all__ = ["BlipImageProcessor"]
+__all__ = ["BlipImageProcessorPil"]
